stickies: edit on type (#3099)
Breakout from https://github.com/tldraw/tldraw/pull/2953
Auto-edit a shape upon keystroke. Related to this change is that we
shouldn't do the selection of all text when editing (and it would
conflict currently with this behavior, or at least we'd have to special
case). In any case, I would argue that it's odd and doesn't feel like
the right move — this PR updates it so that it puts the cursor at the
end of the text which to me feels better or more expected.

### Change Type

- [x] `patch` — Bug fix

### Release Notes

- Sticky note: auto-edit on keystroke.