stickies: edit on type, part 2: place cursor at intended position in general (#3132)
This is what we discussed this morning! Places the cursor at the
intended position. This is doing some magic with opacity to still
capture the cursor position and we show it when we're actually editing.

The discussion offline was that select all, being the current behavior
implies that editing = 'replace all text' as the default behavior,
requiring a third-click to deselect the entirely selected text. The
previous 'edit on type' PR that this is stacked on top of start shifting
that paradigm to at least choose append to text instead of replace -
but! We can still do better. This PR goes the next logical step and
places the cursor where the user would probably expect it to go.

(This obviously also ties in with the textfields PR work in general as
well.)


https://github.com/tldraw/tldraw/assets/469604/7f747296-0801-4c01-baa8-25e33b06882c

### Change Type

<!-- ❗ Please select a 'Scope' label ❗️ -->

- [x] `sdk` — Changes the tldraw SDK
- [ ] `dotcom` — Changes the tldraw.com web app
- [ ] `docs` — Changes to the documentation, examples, or templates.
- [ ] `vs code` — Changes to the vscode plugin
- [ ] `internal` — Does not affect user-facing stuff

<!-- ❗ Please select a 'Type' label ❗️ -->

- [ ] `bugfix` — Bug fix
- [x] `feature` — New feature
- [ ] `improvement` — Improving existing features
- [ ] `chore` — Updating dependencies, other boring stuff
- [ ] `galaxy brain` — Architectural changes
- [ ] `tests` — Changes to any test code
- [ ] `tools` — Changes to infrastructure, CI, internal scripts,
debugging tools, etc.
- [ ] `dunno` — I don't know


### Test Plan

1. Add a step-by-step description of how to test your PR here.
2.

- [ ] Unit Tests
- [ ] End to end tests

### Release Notes

- Add a brief release note for your PR here.