sync: don't auto-create missing rooms; instead show a 'not found' error (#3551)
This PR removes the auto-creation of new rooms via just hitting a URL
endpoint on dotcom. We now serve up (effectively) a 404. The tricky bit
of this is that the 404 status is passed through the websocket. As such,
there's quite a bit of plumbing to be able to bubble up a not-found
event from the server→ client.

I'll copy the comment that I put in the code on how it passes through
the code:
1. From here (`getInitialRoomState`), we send back a `room_not_found` to
`TLDrawDurableObject`.
2. In `TLDrawDurableObject`, we accept and
and then immediately close the client. This lets us send a
`TLCloseEventCode.NOT_FOUND`
  closeCode down to the client.
3. `joinExistingRoom` which handles the websocket upgrade is not
affected.
Again, we accept the connection, it's just that we immediately close
right after.
4. In `ClientWebSocketAdapter`, `ws.onclose` is called, and that calls
`_handleDisconnect`.
5. `_handleDisconnect` sets the status to 'error' and calls the
`onStatusChange` callback.
6. On dotcom in `useRemoteSyncClient`, we have `socket.onStatusChange`
callback
where we set `TLIncompatibilityReason.RoomNotFound` and close the client
+ socket.
7. Finally on the dotcom app we use `StoreErrorScreen` to display an
appropriate msg.

Phew!

<img width="1344" alt="Screenshot 2024-04-23 at 14 13 17"
src="https://github.com/tldraw/tldraw/assets/469604/8f05861f-378a-421c-b53b-a96bd3fa12b8">


### Change Type

<!-- ❗ Please select a 'Scope' label ❗️ -->

- [x] `sdk` — Changes the tldraw SDK
- [ ] `dotcom` — Changes the tldraw.com web app
- [ ] `docs` — Changes to the documentation, examples, or templates.
- [ ] `vs code` — Changes to the vscode plugin
- [ ] `internal` — Does not affect user-facing stuff

<!-- ❗ Please select a 'Type' label ❗️ -->

- [ ] `bugfix` — Bug fix
- [x] `feature` — New feature
- [ ] `improvement` — Improving existing features
- [ ] `chore` — Updating dependencies, other boring stuff
- [ ] `galaxy brain` — Architectural changes
- [ ] `tests` — Changes to any test code
- [ ] `tools` — Changes to infrastructure, CI, internal scripts,
debugging tools, etc.
- [ ] `dunno` — I don't know


### Test Plan

1. Test `/r/random-slug` doesn't create a new room, nor does `/r`
2. Make sure `/r/existing-slug` still works.

### Release Notes

- Rooms: `/r/:roomId` no longer creates a new room. Instead, if the room
doesn't exist we bubble up a `room_not_found` error.